Add files via upload
In this Dashboard I showed an analysis of two important factors of Norther Territory Govt. of Australia while working as a Data Analyst at ABI Partners Ltd. The factors are - 
1. Invoices that are Due 2. Bills that are Payable
The Northern Territory Govt. of Australia uses this Report on a daily basis to know their updated dues and bills so that they can have broad range view of their business growth in terms of their finance management. Most of the time the finance managers use this report.